refactor(test): hoist createGetRequest into shared helpers
